project/requirements.txt
Django==2.0
django-multiselectfield==0.1.8
django-rest-auth==0.9.3
djangorestframework==3.8.2
psycopg2==2.7.4
pytz==2018.4
six==1.11.0
